[CS-3209]: Split UI and logic on TransferCard
